api/requirements.api.txt
pandas==2.2.2
uvicorn==0.30.6
fastapi==0.90.0
joblib==1.4.2
scikit-learn==1.5.1
psutil==6.0.0
PyYAML==6.0.1